OnePlus Nord 2 Review: This Vs Poco F3 GT Vs Realme X7 Max 5G, Over To You Now
The search for the true north is continuing at an even faster pace this year. It hasn’t been too long since we saw the OnePlus Nord CE 5G launched in India and in terms of pricing, that’s there and thereabouts where last year’s OnePlus Nord was priced. Albeit a little lesser, but then, there was also some give and take on the spec sheet. Joining the party is the new OnePlus Nord 2, which now means the Nord smartphone line-up in India will be a very capable troika, at least for the time being. And alongside the OnePlus 9 Pro, OnePlus 9 and the OnePlus 9R, with the OnePlus 8 Pro and the OnePlus 8T still on sale, this is the widest the OnePlus phone line-up has ever been. And if you look carefully, the OnePlus Nord 2 is filling in a very important price band too.